+++ blockchainElements/pythonBackendTests/orderbook.py
@@ -2,6 +2,8 @@ from web3 import Web3
 import json
 import threading
 from time import sleep
+from collections import deque
+from threading import Lock
 
 
 w3 = Web3(Web3.HTTPProvider('http://localhost:8545'))
@@ -16,7 +18,7 @@ TOKEN_ADDRESS = "0x2ef22142186cdbe0d1832cd9ffcbcf02be59dab0"
 
 
 #   transactionHash: "0xad2348a5ff165f0b2ddb1226e86ff21efec08c795685251e998f7aad4ecc40ff"
-CONTRACT_ADDRESS = "0x1610b609fef0287c095ab282d6e769b9a7ade4d8"
+CONTRACT_ADDRESS = "0x9FF73caE2F7dA1200ff948a16A1158189f347CA4"
 CONTRACT_ADDRESS = w3.to_checksum_address(CONTRACT_ADDRESS)
 if w3.is_connected():
     print("Connected to node!")
@@ -33,7 +35,7 @@ with open(KEYSTORE_FILE) as f:
 private_key = w3.eth.account.decrypt(keystore_json, KEYSTORE_PASSWORD)
 account = w3.eth.account.from_key(private_key)
 
-with open("orderbookABI.json") as f:
+with open("updatedOrderBookABI.json") as f:
     contract_abi = json.load(f)
 
 orderbook_contract = w3.eth.contract(address=CONTRACT_ADDRESS, abi=contract_abi)
@@ -68,7 +70,12 @@ def print_orderbook_state():
     print(f"Best Bid: {best_bid}, Best Ask: {best_ask}, Last Price: {last_price}")
 
 
+seen_tx_hashes = deque(maxlen=100)
+tx_lock = Lock()
+
 def handle_event(event):
+    #if event['event'] == 'OrderMatched'"
+
     print(f"\nNew Event: {event['event']}")
     print(f"Tx: {event['transactionHash'].hex()}, Block: {event['blockNumber']}")
     for key, value in event['args'].items():
@@ -76,14 +83,27 @@ def handle_event(event):
             value = value.hex()
         print(f"  {key}: {value}")
 
+
 def listen_for_events(event_name, filters):
     event = getattr(orderbook_contract.events, event_name)
-    event_filter = event.create_filter(from_block='latest', argument_filters=filters)
-    print(f"Listening for {event_name} events with filters {filters}...")
+    last_block = w3.eth.block_number
+    print(f"Listening for {event_name} from block {last_block} with filters {filters}...")
+
     while True:
-        for event in event_filter.get_new_entries():
-            handle_event(event)
+        current_block = w3.eth.block_number
+        #logs = event.get_logs(fromBlock=last_block + 1, toBlock='latest', argument_filters=filters)
+        logs = event.get_logs(from_block=last_block + 1, to_block='latest', argument_filters=filters)
+        
+        for log in logs:
+            tx_hash = log['transactionHash'].hex()
+            with tx_lock:
+                if tx_hash not in seen_tx_hashes:
+                    seen_tx_hashes.append(tx_hash)
+                    handle_event(log)
+
+        last_block = current_block
         sleep(2)
+
 
 def spin_event_threads():
     event_configs = [
@@ -98,11 +118,26 @@ def spin_event_threads():
         thread.start()
 
 
+
 if __name__ == "__main__":
     spin_event_threads()
     print("Event listeners started. Waiting a moment for threads to initialize...")
     sleep(3)  # Give threads a moment to start
 
+    print("\nPlacing a limit buy order: 10 kWh @ 5 tokens/kWh")
+    send_transaction(
+        orderbook_contract.functions.placeOrder(
+            5,  # energyAmount (kWh)
+            5,   # pricePerUnit (tokens/kWh)
+            False,  # isBuy
+            True  # isMarket
+        )
+    )
+
+    sleep(10)
+    print_orderbook_state()
+
+    '''    
     # 2. Place a limit buy order
     print("\nPlacing a limit buy order: 10 kWh @ 5 tokens/kWh")
     send_transaction(
@@ -113,19 +148,24 @@ if __name__ == "__main__":
             False  # isMarket
         )
     )
-
+    
     # 3. Place a limit sell order
     print("\nPlacing a limit sell order: 15 kWh @ 6 tokens/kWh")
     send_transaction(
         orderbook_contract.functions.placeOrder(
-            15,  # energyAmount
-            6,   # pricePerUnit
+            5,  # energyAmount
+            1,   # pricePerUnit
             False,  # isBuy
             False   # isMarket
         )
     )
+    '''
+    #send_transaction(orderbook_contract.functions.placeOrder(10, 4, True, False))
+    #send_transaction(orderbook_contract.functions.placeOrder(8, 3, True, False))
+    #send_transaction(orderbook_contract.functions.placeOrder(5, 2, True, False))
+    #print("\nOrderbook state after limit orders:")
+  
 
-    print("\nOrderbook state after limit orders:")
-    print_orderbook_state()
+
